refactor: move interpolate to util
